use byte indexing, fix op_emit

+++ cpu.h
@@ -1,2 +1,4 @@
+#include "vm.h"
+
 void cpu_run(vm_t *vm);
-unsigned long* operand_pointer(vm_t *vm, int optype, unsigned long* dat);
+unsigned long* operand_pointer(vm_t *vm, int optype, byte* dat);
